# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des utilisateurs et vendeurs.",
        "title": "o'z api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support API",
            "email": "goliyaohugues@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/shops": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère la liste des magasins de l'utilisateur",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Liste tous les magasins",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.shopsResponse"
                        }
                    },
                    "500": {
                        "description": "Erreur de récupération",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ajoute un nouveau magasin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Créer un magasin",
                "parameters": [
                    {
                        "description": "Données du magasin",
                        "name": "shop",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.shopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Shop"
                        }
                    },
                    "400": {
                        "description": "Json Invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shops/assign-vendeur": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lier un vendeur à des magasin.\nNote: Un Admin ne peut assigner que ses propres vendeurs (parent_id doit correspondre).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shops"
                ],
                "summary": "Assigner un vendeur à des magasins",
                "parameters": [
                    {
                        "description": "IDs du vendeur et des magasins",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.AssignShopsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Assignation réussie",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé (Vendeur non lié à cet Admin)",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shops/vendeur": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère tous les magasins associés au vendeur connecté (via son ID d'owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Liste les magasins d'un vendeur",
                "responses": {
                    "200": {
                        "description": "Liste des magasins avec le compte",
                        "schema": {
                            "$ref": "#/definitions/transport.shopsResponse"
                        }
                    },
                    "401": {
                        "description": "Utilisateur non identifié",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Aucun magasin trouvé",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur interne ou JSON invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shops/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère un magasin spécifique par son UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Détails d'un magasin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID du magasin",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Shop"
                        }
                    },
                    "400": {
                        "description": "Format d'identifiant invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Aucun magasin trouvé",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Met à jour les informations d'un magasin existant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Modifier un magasin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID du magasin",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouvelles données",
                        "name": "shop",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.shopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Shop"
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Mise à jour impossible",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Suppression logique (soft delete) d'un magasin",
                "tags": [
                    "Shops"
                ],
                "summary": "Supprimer un magasin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID du magasin",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Echec de suppression",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Réactive un magasin supprimé",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Restaurer un magasin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID du magasin",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Magasin restauré avec succès",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Erreur lors du restore",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/stocks/movements/shop/{shop_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Historique des mouvements pour une boutique donnée",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Mouvements d'une boutique",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la boutique",
                        "name": "shop_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.StockMovement"
                            }
                        }
                    },
                    "400": {
                        "description": "shop_id invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Shop": {
            "type": "object",
            "required": [
                "shopAddress",
                "shopMail",
                "shopName",
                "shopPhone"
            ],
            "properties": {
                "assignedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "integer"
                },
                "shopAddress": {
                    "type": "string"
                },
                "shopID": {
                    "type": "integer"
                },
                "shopMail": {
                    "type": "string"
                },
                "shopName": {
                    "type": "string",
                    "minLength": 3
                },
                "shopPhone": {
                    "type": "string"
                },
                "shopUUID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.StockMovement": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "movement_id": {
                    "type": "integer"
                },
                "movement_type": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity_change": {
                    "type": "integer"
                },
                "shop_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "transport.AssignShopsRequest": {
            "type": "object",
            "required": [
                "shopIDs",
                "vendeurID"
            ],
            "properties": {
                "shopIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "vendeurID": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "transport.shopRequest": {
            "type": "object",
            "required": [
                "shopAddress",
                "shopMail",
                "shopName",
                "shopPhone"
            ],
            "properties": {
                "shopAddress": {
                    "type": "string"
                },
                "shopMail": {
                    "type": "string"
                },
                "shopName": {
                    "type": "string",
                    "minLength": 3
                },
                "shopPhone": {
                    "type": "string"
                }
            }
        },
        "transport.shopsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "shops": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Shop"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" suivi d'un espace et du token JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}